+++ src/views/mod.rs
@@ -160,5 +160,5 @@ pub use table::{Align, Column, Table};
 #[doc(inline)]
 pub use tooltip::Tooltip;
 #[doc(inline)]
-pub use view::{IdleView, View, ViewCore, ViewId};
+pub use view::{View, ViewCore, ViewId};
 pub use window::WindowLike;